HW4.T2

Documentation

data ParseError

Constructors

ErrorAtPos Natural

runP :: Parser a -> String -> Except ParseError a

Applies given parser on given string and returns result

pChar :: Parser Char

Parser that consumes one char. Was given in HW files as an example

parseError :: Parser a

Parser that always throws an error

parseExpr :: String -> Except ParseError Expr

Takes in a string and parses it. Returns the result which is Either Error and tells the position of failure Either Success and the parsed expression.

Production rules used as the grammar:

parserE -> parserT parserE'

parserE' -> + parserT parserE' | - parserT parserE' | epsilon

parserT -> parserF parserT'

parserT' -> * parserF parserT' | / parserF parserT' | epsilon

parserF -> ( parserE ) | int