HW4.T2

Documentation

newtype ParseError

throws in the Parser Monad

Constructors

ErrorAtPos Natural

runP :: Parser a -> String -> Except ParseError a

Runs a computation of the parser

pChar :: Parser Char

Consumes a single char from the string returns an Error if unable.

parseError :: Parser a

Throws in the parser monad

parseExpr :: String -> Except ParseError Expr

The point of entry for our parser of arithmetic expressions.

We are parsing the following CFG that is LL(1); The following are the production rules for the grammar

S -> E

E -> T E'

E' -> + T E' | - T E' | eps

T -> F T'

T' -> * F T' | / F T' | eps

F -> double | ( E )

Examples of use are given below

>>> parseExpr "3.14 + 1.618 * 2"
Success (Op (Add (Val 3.14) (Op (Mul (Val 1.618) (Val 2.0)))))

Can throw an Error if given unexpected input

>>> parseExpr "24 + Hello"
Error (ErrorAtPos 3)